feat(types): mock에 타입 적용
